build: remove the prettier plugin from qlty lint check
The release-please github action lists each change in the change log
starting with a '-'. However, the `qlty check` command raises an error that
fails the build because it expectes unordered lists to be prefixed with '*'.
This is because it uses prettier under the hood which can not be configured
to use any other bullet character.